Scene is flushed continuously, in order to the camera changes to be applied
Scene changes are not send to the renderer before flush() is called. This
ensures that several changes to the scene can be applied at once atomically,
without the scene flichering or tearing. This has the effect that if no
flush is called ever, the camera changes will not be reflected in the
rendered image